Update comments.js.jsx
Hey Justin, Great tutorial! It has really helped me get started using React and Rails. 

One thing: I think the original Facebook tutorial expects all the comments to be returned here so when you return just one Comment, the script is breaking with a *Uncaught TypeError: this.props.data.map is not a function* because it is expecting an array of multiple objects. I only caught this because I removed the polling from my app and noticed that it wasn't working without polling. Also if it did work, then the page would be re-rendered with only the new comment. My solution was just to call loadCommentsFromServer again. Thanks, Andy.